Added support for menus in the windowless window, especially on macOS
