+++ cpp/open3d/visualization/gui/MenuMacOS.h
@@ -32,20 +32,10 @@ namespace open3d {
 namespace visualization {
 namespace gui {
 
-struct DrawContext;
-struct Theme;
-
-/// The menu item action is handled by Window, rather than by registering a
-/// a callback function with (non-existent) Menu::SetOnClicked(). This is
-/// because on macOS the menubar is global over all application windows, so any
-/// callback would need to go find the data object corresponding to the active
-/// window.
-class Menu : public MenuBase {
-    friend class Application;
-
+class MenuMacOS : public MenuBase {
 public:
-    Menu();
-    virtual ~Menu();
+    MenuMacOS();
+    virtual ~MenuMacOS();
 
     void AddItem(const char* name,
                  ItemId item_id = NO_ITEM,
@@ -64,11 +54,7 @@ public:
 
     int GetNumberOfItems() const override;
 
-    /// Searches the menu hierarchy down from this menu to find the item
-    /// and returns true if the item is enabled.
     bool IsEnabled(ItemId item_id) const override;
-    /// Searches the menu hierarchy down from this menu to find the item
-    /// and set it enabled according to \p enabled.
     void SetEnabled(ItemId item_id, bool enabled) override;
 
     bool IsChecked(ItemId item_id) const override;
@@ -76,7 +62,6 @@ public:
 
     int CalcHeight(const Theme& theme) const override;
 
-    /// Returns true if submenu visibility changed on last call to DrawMenuBar
     bool CheckVisibilityChange() const override;
 
     ItemId DrawMenuBar(const DrawContext& context, bool is_enabled) override;
@@ -84,7 +69,7 @@ public:
                 const char* name,
                 bool is_enabled) override;
 
-    void* GetNativePointer() override;  // nullptr if not using native menus
+    void* GetNativePointer() override;
 
 private:
     struct Impl;
