+++ cpp/open3d/visualization/gui/MenuMacOS.mm
@@ -24,9 +24,9 @@
 // IN THE SOFTWARE.
 // ----------------------------------------------------------------------------
 
-#include "open3d/visualization/gui/Menu.h"  // defines GUI_USE_NATIVE_MENUS
+#include "open3d/visualization/gui/MenuMacOS.h"
 
-#if GUI_USE_NATIVE_MENUS
+#if __APPLE__
 
 #import "AppKit/AppKit.h"
 
@@ -60,9 +60,9 @@ namespace open3d {
 namespace visualization {
 namespace gui {
 
-struct Menu::Impl {
+struct MenuMacOS::Impl {
     NSMenu *menu_;
-    std::vector<std::shared_ptr<Menu>> submenus_;
+    std::vector<std::shared_ptr<MenuMacOS>> submenus_;
 
     NSMenuItem* FindMenuItem(ItemId itemId) const {
 
@@ -83,34 +83,34 @@ struct Menu::Impl {
 
 };
 
-Menu::Menu()
-: impl_(new Menu::Impl()) {
+MenuMacOS::MenuMacOS()
+: impl_(new MenuMacOS::Impl()) {
     impl_->menu_ = [[NSMenu alloc] initWithTitle:@""];
     impl_->menu_.autoenablesItems = NO;
 }
 
-Menu::~Menu() {} // ARC will automatically release impl_->menu
+MenuMacOS::~MenuMacOS() {} // ARC will automatically release impl_->menu
 
-void* Menu::GetNativePointer() { return impl_->menu_; }
+void* MenuMacOS::GetNativePointer() { return impl_->menu_; }
 
-void Menu::AddItem(const char *name,
-                   ItemId item_id /*= NO_ITEM*/,
-                   KeyName key /*= KEY_NONE*/) {
+void MenuMacOS::AddItem(const char *name,
+                        ItemId item_id /*= NO_ITEM*/,
+                        KeyName key /*= KEY_NONE*/) {
     InsertItem(impl_->menu_.numberOfItems, name, item_id, key);
 }
 
-void Menu::AddMenu(const char *name, std::shared_ptr<Menu> submenu) {
+void MenuMacOS::AddMenu(const char *name, std::shared_ptr<MenuBase> submenu) {
     InsertMenu(impl_->menu_.numberOfItems, name, submenu);
 }
 
-void Menu::AddSeparator() {
+void MenuMacOS::AddSeparator() {
     [impl_->menu_ addItem: [NSMenuItem separatorItem]];
 }
 
-void Menu::InsertItem(int index,
-                      const char* name,
-                      ItemId item_id /*= NO_ITEM*/,
-                      KeyName key /*= KEY_NONE*/) {
+void MenuMacOS::InsertItem(int index,
+                           const char* name,
+                           ItemId item_id /*= NO_ITEM*/,
+                           KeyName key /*= KEY_NONE*/) {
     std::string shortcut;
     shortcut += char(key);
     NSString *objc_shortcut = [NSString stringWithUTF8String:shortcut.c_str()];
@@ -129,9 +129,10 @@ void Menu::InsertItem(int index,
     }
 }
 
-void Menu::InsertMenu(int index, const char* name,
-                      std::shared_ptr<Menu> submenu) {
-    submenu->impl_->menu_.title = [NSString stringWithUTF8String:name];
+void MenuMacOS::InsertMenu(int index, const char* name,
+                           std::shared_ptr<MenuBase> submenu) {
+    auto macos_submenu = std::dynamic_pointer_cast<MenuMacOS>(submenu);
+    macos_submenu->impl_->menu_.title = [NSString stringWithUTF8String:name];
     auto item = [[NSMenuItem alloc]
                  initWithTitle:[NSString stringWithUTF8String:name]
                         action:nil
@@ -141,19 +142,19 @@ void Menu::InsertMenu(int index, const char* name,
     } else {
         [impl_->menu_ addItem:item];
     }
-    [impl_->menu_ setSubmenu:submenu->impl_->menu_ forItem:item];
-    impl_->submenus_.insert(impl_->submenus_.begin() + index, submenu);
+    [impl_->menu_ setSubmenu:macos_submenu->impl_->menu_ forItem:item];
+    impl_->submenus_.insert(impl_->submenus_.begin() + index, macos_submenu);
 }
 
-void Menu::InsertSeparator(int index) {
+void MenuMacOS::InsertSeparator(int index) {
     [impl_->menu_ insertItem: [NSMenuItem separatorItem] atIndex: index];
 }
 
-int Menu::GetNumberOfItems() const {
+int MenuMacOS::GetNumberOfItems() const {
     return impl_->menu_.numberOfItems;
 }
 
-bool Menu::IsEnabled(ItemId item_id) const {
+bool MenuMacOS::IsEnabled(ItemId item_id) const {
     NSMenuItem *item = impl_->FindMenuItem(item_id);
     if (item) {
         return (item.enabled == YES ? true : false);
@@ -161,14 +162,14 @@ bool Menu::IsEnabled(ItemId item_id) const {
     return false;
 }
 
-void Menu::SetEnabled(ItemId item_id, bool enabled) {
+void MenuMacOS::SetEnabled(ItemId item_id, bool enabled) {
     NSMenuItem *item = impl_->FindMenuItem(item_id);
     if (item) {
         item.enabled = (enabled ? YES : NO);
     }
 }
 
-bool Menu::IsChecked(ItemId item_id) const {
+bool MenuMacOS::IsChecked(ItemId item_id) const {
     NSMenuItem *item = impl_->FindMenuItem(item_id);
     if (item) {
         return (item.state == NSControlStateValueOn);
@@ -176,7 +177,7 @@ bool Menu::IsChecked(ItemId item_id) const {
     return false;
 }
 
-void Menu::SetChecked(ItemId item_id, bool checked) {
+void MenuMacOS::SetChecked(ItemId item_id, bool checked) {
     NSMenuItem *item = impl_->FindMenuItem(item_id);
     if (item) {
         item.state = (checked ? NSControlStateValueOn
@@ -184,21 +185,22 @@ void Menu::SetChecked(ItemId item_id, bool checked) {
     }
 }
 
-int Menu::CalcHeight(const Theme &theme) const {
+int MenuMacOS::CalcHeight(const Theme &theme) const {
     return 0;  // menu is not part of window on macOS
 }
 
-bool Menu::CheckVisibilityChange() const {
+bool MenuMacOS::CheckVisibilityChange() const {
     return false;
 }
 
-Menu::ItemId Menu::DrawMenuBar(const DrawContext &context, bool is_enabled) {
+MenuBase::ItemId MenuMacOS::DrawMenuBar(const DrawContext &context,
+                                        bool is_enabled) {
     return NO_ITEM;
 }
 
-Menu::ItemId Menu::Draw(const DrawContext &context,
-                        const char *name,
-                        bool is_enabled) {
+MenuBase::ItemId MenuMacOS::Draw(const DrawContext &context,
+                                 const char *name,
+                                 bool is_enabled) {
     return NO_ITEM;
 }
 
@@ -206,4 +208,4 @@ Menu::ItemId Menu::Draw(const DrawContext &context,
 }  // namespace visualization
 }  // namespace open3d
 
-#endif  // GUI_USE_NATIVE_MENUS
+#endif  // __APPLE__
